Add shared base class for HTML builder adapters
We can share some private methods between `FromLiquid` and `FromERB`. We
can also use the base class as documentation for future adapter
implementations.
